introduce translation feature
this commit introduces the translation feature powered by the translation api of MLKit. For now, a custom bottomsheetdialog.xml layout is provided, which automatically triggers the translation of recognized objects and shows it under the original detected label. For now, the only language supported is italian, but we'll need to provide the possibility to choose the language.

Signed-off-by: readbeard <matteo.principe92@gmail.com>